Cleanup & refactoring
