+++ deletefb/deletefb.py
@@ -1,16 +1,16 @@
 #!/usr/bin/env python
 
+from .tools.common import logger
+from .tools.config import settings
+from .tools.likes import unlike_pages
+from .tools.login import login
+from .tools.wall import delete_posts
+
 import argparse
 import getpass
 import json
 import os
 import sys
-
-from .tools.config import settings
-from .tools.common import logger
-from .tools.login import login
-from .tools.wall import delete_posts
-from .tools.likes import unlike_pages
 
 LOG = logger("deletefb")
