+++ deletefb/tools/wall.py
@@ -1,9 +1,27 @@
-import time
+from .archive import archiver, Post
+from .common import SELENIUM_EXCEPTIONS, click_button
+from .config import settings
 from selenium.webdriver.common.action_chains import ActionChains
 
-from .config import settings
-from .common import SELENIUM_EXCEPTIONS, click_button
-from .archive import archiver, Post
+import attr
+import datetime
+import time
+import uuid
+
+# Data type definitions of posts and comments
+@attr.s
+class Post:
+    content = attr.ib()
+    comments = attr.ib(default=[])
+    date = attr.ib(factory=datetime.datetime.now)
+    name = attr.ib(factory=lambda: uuid.uuid4().hex)
+
+@attr.s
+class Comment:
+    commenter = attr.ib()
+    content = attr.ib()
+    date = attr.ib(factory=datetime.datetime.now)
+    name = attr.ib(factory=lambda: uuid.uuid4().hex)
 
 # Used as a threshold to avoid running forever
 MAX_POSTS = settings["MAX_POSTS"]
